move stream id space to large unsigned ints
- moved from short unsigned to unsigned
- internal and user stream ids start at 2^30,
  leaves a billion ids for BE singlecast streams
- instead of directly referring to 0 for ctl
  stream, 1 for topol stream, etc., we now
  define some constants (e.g., CTL_STRM_ID)

+++ lightweight/src/Filter.c
@@ -16,7 +16,7 @@
  * nodes, so these functions are in place only as placeholders
  * if we choose to implement the functionality at a later time. */
 
-Filter_t* new_Filter_t(unsigned short iid)
+Filter_t* new_Filter_t(unsigned int iid)
 {
     return NULL;
 }
